Revert "[PNI] Update feature branch (#11383)"
This reverts commit 9222eb17cb49a1f30ce65fcb14fcecb554a0e827.
