+++ dev-requirements.txt
@@ -74,7 +74,7 @@ jsbeautifier==1.14.7
     #   djlint
 mccabe==0.7.0
     # via flake8
-mypy==1.6.1
+mypy==1.5.1
     # via -r dev-requirements.in
 mypy-extensions==1.0.0
     # via
