--- network-api/networkapi/donate/factory/customblocks/__init__.py
@@ -1,2 +1,0 @@
-# flake8: noqa
-from .notice_block import NoticeBlockFactory
