--- network-api/networkapi/donate/pagemodels/base_page.py
@@ -1,8 +1,0 @@
-from wagtail.models import Page
-
-from networkapi.wagtailpages.models import FoundationMetadataPageMixin
-
-
-class BaseDonationPage(FoundationMetadataPageMixin, Page):
-    class Meta:
-        abstract = True
