+++ network-api/networkapi/settings.py
@@ -264,7 +264,6 @@ INSTALLED_APPS = list(
             "networkapi.wagtailpages",
             "networkapi.mozfest",
             "networkapi.donate",
-            "networkapi.donate_banner",
             "networkapi.reports",
             "pattern_library" if PATTERN_LIBRARY_ENABLED else None,
             "networkapi.project_styleguide",
