+++ network-api/networkapi/templates/buyersguide/fragments/consumer_creepometer_page_2023/best_and_worst_products.html
@@ -15,7 +15,7 @@
           <h3 class="{{ h3_class }} tw-col-span-2 medium:tw-col-span-5 tw-text-center tw-mb-8" id="winners-grid-title">The Winners</h3>
           {% include grid_item_template with body_small_class=body_small_class product_name_class=product_name_class item_class="tw-col-span-2 medium:tw-col-span-1" company="Technology Foundation" product="Signal" img_url="_images/buyers-guide/consumer-creepometer/products/Signal.png" link="https://foundation.mozilla.org/privacynotincluded/signal/" %}
           {% include grid_item_template with body_small_class=body_small_class product_name_class=product_name_class item_class="tw-col-span-1" company="Sonos" product="SL Speakers" img_url="_images/buyers-guide/consumer-creepometer/products/Sonos_SL_Speakers.png" link="https://foundation.mozilla.org/privacynotincluded/sonos-sl-speakers/" %}
-          {% include grid_item_template with body_small_class=body_small_class product_name_class=product_name_class item_class="tw-col-span-1" company="Touchkin" product="Wysa" img_url="_images/buyers-guide/consumer-creepometer/products/Wysa.png" link="https://foundation.mozilla.org/privacynotincluded/wysa/" %}
+          {% include grid_item_template with body_small_class=body_small_class product_name_class=product_name_class item_class="tw-col-span-1" company="Touchkin" product="Wysa" img_url="_images/buyers-guide/consumer-creepometer/products/Wysa.png" link="https://foundation.mozilla.org/en/privacynotincluded/wysa/" %}
           {% include grid_item_template with body_small_class=body_small_class product_name_class=product_name_class item_class="tw-col-span-1" company="Pocketbook" product="Pocketbook" img_url="_images/buyers-guide/consumer-creepometer/products/Pocketbook.png" link="https://foundation.mozilla.org/privacynotincluded/pocketbook/" %}
           {% include grid_item_template with body_small_class=body_small_class product_name_class=product_name_class item_class="tw-col-span-1" company="Garmin" product="Garmin Fenix" img_url="_images/buyers-guide/consumer-creepometer/products/Garmin_Fenix.png" link="https://foundation.mozilla.org/privacynotincluded/garmin-fenix/" %}
         </div>
