+++ network-api/networkapi/templates/buyersguide/fragments/consumer_creepometer_page_2023/hero.html
@@ -12,7 +12,7 @@
       <div class="tw-w-full medium:tw-w-[82.5%]">
         <h1 class="tw-font-normal tw-font-zilla tw-text-[48px] tw-leading-[44px] medium:tw-text-[80px] medium:tw-leading-none tw-mb-12 medium:tw-mb-24">Welcome to <span class="tw-font-semibold">Mozilla’s</span> first-ever <strong class="tw-font-semibold">Annual Consumer Creep-O-Meter</strong></h1>
         <h2 class="tw-font-light tw-font-zilla tw-text-[24px] tw-leading-[28px] medium:tw-text-[40px] medium:tw-leading-[48px]">
-          In 2023, the state of our digital privacy is <strong class="tw-relative tw-block tw-text-center tw-capitalize tw-font-sans tw-font-[28px] tw-leading-[32px] medium:tw-text-[80px] medium:tw-leading-none tw-font-extrabold tw-max-w-[65%] small:tw-max-w-[75%] tw-my-12 medium:tw-my-24 tw-px-12 tw-py-6 medium:tw-py-12 tw-bg-blue-05 tw-rounded-lg medium:tw-rounded-3xl after:tw-content-[''] after:tw-w-[calc(94px/2)] after:tw-h-[calc(57px/2)] medium:after:tw-w-[94px] medium:after:tw-h-[57px] after:tw-absolute after:tw-top-[calc(100%-2px)] after:tw-right-0 after:tw-bg-center after:tw-bg-no-repeat after:tw-bg-cover after:tw-bg-[url('../_images/buyers-guide/consumer-creepometer/bubble-stem.svg')]">very creepy</strong>
+          In 2023, the state of our digital privacy is <strong class="tw-relative tw-block tw-text-center tw-capitalize tw-font-sans tw-font-[28px] tw-leading-[32px] medium:tw-text-[80px] medium:tw-leading-none tw-font-extrabold tw-max-w-[75%] tw-my-12 medium:tw-my-24 tw-px-12 tw-py-6 medium:tw-py-12 tw-bg-blue-05 tw-rounded-lg medium:tw-rounded-3xl after:tw-content-[''] after:tw-w-[calc(94px/2)] after:tw-h-[calc(57px/2)] medium:after:tw-w-[94px] medium:after:tw-h-[57px] after:tw-absolute after:tw-top-[calc(100%-2px)] after:tw-right-0 after:tw-bg-center after:tw-bg-no-repeat after:tw-bg-cover after:tw-bg-[url('../_images/buyers-guide/consumer-creepometer/bubble-stem.svg')]">very creepy</strong>
         </h2>
       </div>
     </div>
