+++ network-api/networkapi/templates/buyersguide/fragments/consumer_creepometer_page_2023/outro.html
@@ -7,9 +7,9 @@
       <div class="tw-col-span-2 tw-p-14 medium:tw-p-24 tw-bg-white tw-rounded-2xl">
         <h2 class="{{ h2_class }}">Our Approach</h2>
         {% with p_class="tw-text-gray-80" %}
-          <p class="{{ p_class }}">The Annual Consumer Creep-O-Meter is calculated using both quantitative and qualitative measures. Our researchers average together findings from our 2023 and late 2022 *Privacy Not Included editions, each of which entails hundreds of hours of research. In 2023 and late 2022, we published three editions and <a class="tw-underline {{ p_class }}" href="https://foundation.mozilla.org/blog/mozilla-study-data-privacy-labels-for-most-top-apps-in-google-play-store-are-false-or-misleading/" target="_blank">one research report</a> covering 172 different products, from cars and mental health apps to fitness trackers and kids toys.</p>
+          <p class="{{ p_class }}">The Annual Consumer Creep-O-Meter is calculated using both quantitative and qualitative measures. Our researchers average together findings from our 2023 and late 2022 *Privacy Not Included editions, each of which entails hundreds of hours of research. In 2023 and late 2022, we published three editions and <a href="https://foundation.mozilla.org/blog/mozilla-study-data-privacy-labels-for-most-top-apps-in-google-play-store-are-false-or-misleading/" target="_blank">one research report</a> covering 172 different products, from cars and mental health apps to fitness trackers and kids toys.</p>
           <p class="{{ p_class }}">Our researchers also assess broader industry trends and developments, like privacy legislation and enforcement, factoring them into the score.</p>
-          <p class="{{ p_class }}">Learn more about <a class="tw-underline {{ p_class }}" href="{{ methodology_url }}" target="_blank">our methodology</a>.</p>
+          <p class="{{ p_class }}">Learn more about <a class="tw-underline {{ p_class }}" href="{{ methodology_url }}">our methodology</a>.</p>
         {% endwith %}
       </div>
     </div>
