+++ network-api/networkapi/templates/buyersguide/fragments/consumer_creepometer_page_2023/stay_in_control.html
@@ -23,7 +23,7 @@
       </div>
       <div class="{{ item_class }} tw-bg-right-top large:tw-bg-left-top tw-bg-[64px_auto] large:tw-bg-[98px_auto] tw-bg-[url('../_images/buyers-guide/consumer-creepometer/money.png')]">
         <h4 class="{{ heading_class }} tw-w-3/4 small:tw-w-full large:tw-mt-64 tw-mt-4 small:tw-mt-12 ">Wield Your Pocketbook</h4>
-        <p class="{{ description_class }} tw-mb-0">If you’re not comfortable with a product’s privacy, don’t buy it. And, speak up. Over the years, we’ve seen companies respond to consumer demand for privacy, like when <a class="tw-underline tw-text-gray-80" href="https://foundation.mozilla.org/blog/applause-for-apples-idfa-decision/" target="_blank">Apple</a> reformed app tracking and <a class="tw-underline tw-text-gray-80" href="https://foundation.mozilla.org/blog/zoom-makes-right-call-encryption/" target="_blank">Zoom</a> made end-to-end encryption a free feature.</p>
+        <p class="{{ description_class }} tw-mb-0">If you’re not comfortable with a product’s privacy, don’t buy it. And, speak up. Over the years, we’ve seen companies respond to consumer demand for privacy, like when Apple reformed app tracking and Zoom made end-to-end encryption a free feature.</p>
       </div>
     </div>
   {% endwith %}
