+++ network-api/networkapi/templates/buyersguide/fragments/consumer_creepometer_page_2023/sub_nav.html
@@ -7,7 +7,7 @@
     </div>
     <h2 class="{{ h2_class }} tw-w-5/6 tw-mb-12 medium:tw-mb-20">Mozilla’s Annual Consumer Creep-O-Meter distills what’s good, what’s bad, and what’s just plain creepy in the world of consumer tech.</h2>
     <div class="tw-w-full medium:tw-w-4/6 tw-mb-20 medium:tw-mb-24">
-      <p class="medium:tw-text-[20px] medium:tw-leading-[30px] tw-font-light tw-mb-12">Since 2017, Mozilla has published 15 editions of <a class="tw-text-black tw-underline hover:tw-text-black hover:tw-underline" href="https://foundation.mozilla.org/privacynotincluded/" target="_blank">*Privacy Not Included</a>, our consumer tech buyers guide. We’ve reviewed over 500 gadgets, apps, cars, and more, assessing their security features, what data they collect, and who they share that data with.</p>
+      <p class="medium:tw-text-[20px] medium:tw-leading-[30px] tw-font-light tw-mb-12">Since 2017, Mozilla has published 15 editions of <a class="tw-text-black tw-underline hover:tw-text-black hover:tw-underline" href="https://foundation.mozilla.org/en/privacynotincluded/" target="_blank">*Privacy Not Included</a>, our consumer tech buyers guide. We’ve reviewed over 500 gadgets, apps, cars, and more, assessing their security features, what data they collect, and who they share that data with.</p>
       <p class="medium:tw-text-[20px] medium:tw-leading-[30px] tw-font-light">In 2023, we compared our most recent findings with those of the past five years. It quickly became clear that products and companies are collecting more personal data than ever before — and then using that information in shady ways.</p>
     </div>
     <nav role="navigation" class="tw-flex tw-flex-col large:tw-flex-row">
