--- network-api/networkapi/templates/donate/pages/help_page.html
@@ -1,35 +1,0 @@
-{% extends "../pages/base.html" %}
-{% load static i18n l10n wagtailcore_tags wagtailimages_tags %}
-
-{% block body_id %}help{% endblock body_id %}
-
-{% block extended_head %}
-    {% include "../fragments/formassembly_head.html" %}
-{% endblock extended_head %}
-
-{% block content %}
-    <div class="tw-container tw-py-10">
-        <div class="cms donate-help-page-content">
-            <h1 class="tw-h1-heading"> {{ page.title }} </h1>
-            {% if page.notice %}
-                {% for block in page.notice %}
-                    {% include_block block %}
-                {% endfor %}
-            {% endif %}
-            <p>
-                {% blocktrans trimmed %}
-                    If you need help with a <a href="https://foundation.mozilla.org/?form=donate&utm_medium=web&utm_source=faq-page&utm_campaign=23-Donor-Care" target="_blank">donation</a> to the Mozilla Foundation, please select the reason for your inquiry from the drop down and a donor care representative will get back to you as soon as possible.
-                {% endblocktrans %}
-            </p>
-            <div class="tw-border-t tw-border-b tw-py-10 tw-mt-28 tw-mb-10 tw-border-gray-20">
-                <p class="tw-h3-heading"> Contact Us </p>
-                {% include "../fragments/formassembly_body.html" %}
-                <link rel="stylesheet" href="{% static "_css/formassembly-override.compiled.css" %}">
-            </div>
-            {% for block in page.body %}
-                {% include_block block with parent_page=page %}
-            {% endfor %}
-        </div>
-    </div>
-{% endblock content %}
-
