+++ network-api/networkapi/templates/pages/base.html
@@ -90,8 +90,8 @@
         {% endif %}
 
         {% block donate_banner %}
-            {% if settings.donate_banner.SiteDonateBanner.active_donate_banner %}
-                {% include "partials/donate_banner.html" with banner=settings.donate_banner.SiteDonateBanner.active_donate_banner.localized %}
+            {% if settings.wagtailpages.FeatureFlags.activate_donate_banner %}
+                {% include "partials/banner.html" %}
             {% endif %}
         {% endblock %}
