--- network-api/networkapi/templates/partials/donate_banner.html
@@ -1,45 +1,0 @@
-{% load settings_value i18n static wagtailimages_tags %}
-{% image banner.background_image fill-1305x720 as background_image %}
-
-{% with wrapper_class="tw-py-24 tw-shadow-[0_0_3px_2px_rgba(9,32,77,.12)_inset]" background_class="tw-bg-[100%_0] tw-bg-no-repeat tw-bg-cover medium:tw-bg-contain" btn_class="tw-text-white tw-border-black tw-bg-blue-40 hover:tw-bg-blue-80 tw-shadow-black" %}
-    <div class="donate-banner {{ background_class }} {{ wrapper_class }} tw-relative tw-z-20 tw-w-full tw-hidden print:tw-hidden" style="background-image:url({{ background_image.url }});">
-        <a href="#"
-           class="tw-p-8
-                  tw-box-content
-                  banner-close
-                  tw-cursor-pointer
-                  tw-absolute
-                  tw-right-0
-                  tw-top-0
-                  tw-group" aria-label="{% trans "Close Banner" %}">
-            <img src="{% static "_images/x.svg" %}" width="24" height="23" alt="{% trans "Close Banner" %}" class="tw-w-8 tw-h-8 tw-transition-transform tw-duration-100 tw-ease-in-out tw-transform group-hover:tw-scale-110 tw-filter tw-invert">
-        </a>
-
-        <div class="tw-container">
-            <div class="tw-row tw-px-6">
-                {% block left_banner_column %}
-                    <div class="tw-px-12 tw-py-16 tw-bg-[hsla(0,0%,100%,.7)] tw-backdrop-blur-sm large:tw-w-8/12 xlarge:tw-w-6/12">
-                        <p class="tw-h2-heading tw-mb-0 tw-text-blue-80 tw-font-semibold tw-italic">
-                            {{ banner.title }}
-                        </p>
-                        <p class="tw-body-large tw-mt-8 tw-leading-normal">
-                            {{ banner.subtitle }}
-                        </p>
-                        <p class="tw-body-large">
-                            {{ banner.cta_text }}
-                        </p>
-                        <div class="tw-w-full tw-pt-6">
-                            <a class="tw-btn-pop {{ btn_class }} tw-shadow-[4px_4px]" tw-w-full tw-text-base active:tw-text-white active:tw-bg-[#0a0c8f] active:tw-border-black" href="{{ banner.cta_link }}">
-                                                                                                                                                                                                             {{ banner.cta_button_text }}
-                                                                                                                                                                                                             </a>
-                                                                                                                                                                                                             </div>
-                                                                                                                                                                                                             </div>
-                                                                                                                                                                                                 {% endblock left_banner_column %}
-
-                                                                                                                                                                                                 {% block right_banner_column %}
-                                                                                                                                                                                                 {% endblock right_banner_column %}
-
-                                                                                                                                                                                                 </div>
-                                                                                                                                                                                                 </div>
-                                                                                                                                                                                                 </div>
-                                                                                                                                                                                 {% endwith %}
