+++ network-api/networkapi/wagtailpages/factory/publication.py
@@ -18,8 +18,6 @@ article_body_streamfield_fields = [
     "callout",
     "content",
     "full_width_image",
-    "image_grid",
-    "image_text",
 ]
