+++ network-api/networkapi/wagtailpages/models.py
@@ -64,6 +64,7 @@ from .pagemodels.campaigns import (
     Signup,
 )
 from .pagemodels.dear_internet import DearInternetPage
+from .pagemodels.feature_flags.feature_flags import FeatureFlags
 from .pagemodels.index import IndexPage
 from .pagemodels.libraries.detail_page import LibraryDetailLinkBase, LibraryDetailPage
 from .pagemodels.libraries.rcc.authors_index import RCCAuthorsIndexPage
