+++ network-api/networkapi/wagtailpages/templates/wagtailpages/pages/youtube-regrets-2021/fragments/intro.html
@@ -3,7 +3,7 @@
     <div class="row justify-content-center">
         <div class="col-12 col-md-8">
             <p>
-                {% blocktrans with stories_link='class="tw-font-bold" href="https://foundation.mozilla.org/youtube/regrets/"' trimmed %}
+                {% blocktrans with stories_link='class="tw-font-bold" href="https://foundation.mozilla.org/campaigns/youtube-regrets/"' trimmed %}
                     In 2019, Mozilla collected <a {{ stories_link }}>countless stories</a> from people whose lives were impacted by YouTube’s recommendation algorithm. People had been exposed to misinformation, developed unhealthy body images, and became trapped in bizarre rabbit holes. The more stories we read, the more we realized how central YouTube has become to the lives of so many people — and just how much YouTube recommendations can impact their wellbeing.
                 {% endblocktrans %}
             </p>
