+++ package.json
@@ -17,7 +17,7 @@
     "build:css": "run-p build:css:file:**",
     "build:css:file:main": "sass source/sass/main.scss network-api/networkapi/temp/frontend/_css/main.compiled.css",
     "build:css:file:bg": "sass source/sass/buyers-guide/bg-main.scss network-api/networkapi/temp/frontend/_css/buyers-guide.compiled.css",
-    "build:css:file:donate": "sass source/sass/donate/donate-main.scss network-api/networkapi/temp/frontend/_css/donate.compiled.css",
+    "build:css:file:donate": "sass source/sass/donate.scss network-api/networkapi/temp/frontend/_css/donate.compiled.css",
     "build:css:file:formassembly": "sass source/sass/formassembly-override.scss network-api/networkapi/temp/frontend/_css/formassembly-override.compiled.css",
     "docker:up": "docker-compose up",
     "docker:down": "docker-compose down",
