+++ pyproject.toml
@@ -15,7 +15,6 @@ ignore="H017"
 [tool.djlint.per-file-ignores]
 "maintenance/maintenance.html" = "D004,H005,H006,H013,H014,H025,H026,H030,H031"
 "network-api/networkapi/templates/donate/fragments/footer.html" = "T003"
-"network-api/networkapi/templates/donate/fragments/formassembly_body.html" = "H020,H021,H026"
 "network-api/networkapi/templates/donate/pages/base.html" = "T003"
 "network-api/networkapi/templates/donate/pages/landing_page.html" = "H006, T003"
 "network-api/networkapi/mozfest/templates/fragments/event_card.html" = "H006"
@@ -100,6 +99,7 @@ ignore="H017"
 "network-api/networkapi/templates/pages/publication_page.html" = "T003,T032"
 "network-api/networkapi/templates/pages/styleguide.html" = "H006,H013,H014,H026,T003,T032"
 "network-api/networkapi/templates/partials/404.html" = "D018,H006"
+"network-api/networkapi/templates/partials/banner.html" = "H006,H025,T003"
 "network-api/networkapi/templates/partials/footer.html" = "H006,H008,T002,T003"
 "network-api/networkapi/templates/partials/primary_nav.html" = "D018,H014,H025,T002,T003"
 "network-api/networkapi/templates/pattern_library_base.html" = "H030,H031,T003"
