+++ requirements.txt
@@ -208,7 +208,7 @@ s3transfer==0.6.1
     # via boto3
 scout-apm==2.26.1
     # via -r requirements.in
-sentry-sdk==1.32.0
+sentry-sdk==1.30.0
     # via -r requirements.in
 six==1.16.0
     # via
