+++ source/sass/donate.scss
@@ -1,0 +1,30 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+*,
+::before,
+::after {
+  box-sizing: border-box;
+}
+
+article,
+aside,
+figcaption,
+figure,
+footer,
+header,
+hgroup,
+main,
+nav,
+section {
+  display: block;
+}
+
+img {
+  max-width: 100%;
+}
+
+#view-landing .intro .rich-text {
+  @apply medium:tw-body-large;
+}
