--- source/sass/donate/blocks/notice-block.scss
@@ -1,11 +1,0 @@
-.notice-block-text {
-  .rich-text {
-    p {
-      margin-bottom: 0;
-
-      &:not(:first-child) {
-        margin-top: 0.25rem;
-      }
-    }
-  }
-}
