--- source/sass/donate/donate-main.scss
@@ -1,53 +1,0 @@
-// mofo-bootstrap
-@import "../mofo-bootstrap/mofo-bootstrap";
-
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-// Custom variables
-@import "../variables";
-
-// Misc
-@import "../cms";
-
-// Wagtail Stream Blocks
-@import "./blocks/notice-block.scss";
-@import "../wagtail/blocks/airtable";
-@import "../wagtail/blocks/feature-quote";
-@import "../wagtail/blocks/image-text-mini";
-@import "../wagtail/blocks/iframe-block";
-@import "../wagtail/blocks/link-button.scss";
-@import "../wagtail/blocks/rich-text";
-@import "../wagtail/blocks/profiles";
-@import "../wagtail/blocks/video-block";
-@import "../wagtail/blocks/article-blocks";
-@import "../wagtail/blocks/image-feature";
-@import "../wagtail/blocks";
-
-*,
-::before,
-::after {
-  box-sizing: border-box;
-}
-
-article,
-aside,
-figcaption,
-figure,
-footer,
-header,
-hgroup,
-main,
-nav,
-section {
-  display: block;
-}
-
-img {
-  max-width: 100%;
-}
-
-#view-landing .intro .rich-text {
-  @apply medium:tw-body-large;
-}
