+++ source/sass/wagtail/blocks.scss
@@ -69,13 +69,6 @@
   }
 }
 
-.donate-help-page-content {
-  .streamfield-content {
-    @extend .px-0;
-    @extend .col-12;
-  }
-}
-
 .mozfest-content {
   &.two-col {
     @extend .p-0;
